pyproject.toml
[tool.poetry]
name = "sso"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
flask = "^3.0.2"
pysaml2 = "^7.5.0"
python-dotenv = "^1.0.1"
requests-oauthlib = "^2.0.0"
gunicorn = "^22.0.0"
httpx = "^0.27.0"
flask-session = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
